ugc_flask/requirements.txt
alembic==1.13.2
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
pika==1.3.2
psycopg2-binary==2.9.8
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
SQLAlchemy==2.0.34
sqlparse==0.5.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.4
